requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.30.0
anthropic>=0.25.0
customtkinter>=5.2.2
pillow>=10.3.0
